requirements.txt
Pillow==7.1.2
requests==2.23.0
tqdm==4.46.0